don't use writer and state monad
we can track the state of the scoped by using a module with a ref
inside. By providing the interface we might then also be able to swap it
out for traced semantics.
Then all remaining GenM functions only use Reader and Error so we can
move them to the others. So GenM can be at the bottom of it's module.
Still need to remove all the cruft from the GenM monad though