use Base instead of Core because the latter conflicts with coq
This seems to be pretty shit in OCaml. Apparenly two libraries cannot
define the same module (Flags in my case) because a library is a flat
collection of "compilation units" (files), and all compilation units in
all libraries must be named differently.
This
answer (https://stackoverflow.com/questions/42351018/in-ocaml-two-third-party-libraries-expose-same-module-name-failure-to-link)
suggests that the only fix is that upstream renames one of the files.
This seems to be a horrifying aspect about the ocaml ecosystem.