removing term datatype.
It should be possible if I just remove the SubstTerm constructor so that
SubstTy cannot be terms anymore. Then I just need to transform them into
terms while generating the code instead of when translating to coq AST.
And then I could also completely remove the custom AST because
translation is now 1 to 1.