add support for unscoped syntax and refactor a bit
In Haskell the support for unscoped code was during the translation from
a custom AST to strings representing coq code. We do it here while
constructing the Coq AST. At the moment the information whether we use
unscoped/scoped code is saved in a global variable. I could also have
added it to the reader monad but that would have meant more refactoring
due to bad composability with higher order functions.