fix some bugs while trying to compile pi calculus
I had mixed up patternSId and patternSIdNoRen.
Monadic code unfortunately makes debugging harder as I was looking in
the wrong function before wondering why execution suddenly stopped. It
was because the first let* suspends computation and in the end the
monadic result of the function was never forced so my printf-debug code
was never executed.