totally hack together printing of arrows
For some reason a less hacky version of the setup_notation function
works when I call it from the repl, but not when I execute the copiled
program. Might be a bug, or differences in native vs byte compiled
code (provided dune used byte compiled version of code).

Basically we first define a scope and then add a notation to it. In the
repl we can set the level of the notation to 70. In compiled code I had
t ouse 1000. Maybe I also need to add associativity (atm it also prints
it fully parenthesized)