create some of the automation
I'm having some doubts whether this is the easiest way. The Haskell
implementation just derived all the information (which lemmas belong to
asimpl etc) again, whereas I register a lemma for asimpl when I generate
it during code generation.
Conceptually I find my approach better because it should better handle
refactorings and it's pretty easy to insert a statement during
code generation to register something that belongs to the automation.
But the disadvantage is that it not centralized. The information about
which automation is generated flows from varous parts of the code
generation.
Anyways my approach can easily be reworked to the other one, I just need
to fill the `info` variable centrally instead of in various places.