fix Kathrin's case study a bit
by comparing with state of ACP submission.
I had changed the static files a bit too much and fsimpl was too weak.
Unfortunately the strong fsimpl also uses functional extensionality.
Now most lemmas go through with my generated tactics. Some are still not
working but by comparing with the files submitted during ACP I should be
able to find out what's wrong there.
The order in which notations are declared seems to matter (most are
only printing so it should not have behaviour on the tactics).