try out the under tactic
If our goal is of the form `s[t] = ?e` we don't want to rewrite on the
right side, but Coq's rewriting tactic sometimes does it because the
right side it too malleable. The under tactic abstracts the goal and
only allows rewriting on the left-hand side. We could either use it
directly (and thus ssreflect) or reimplement the functionality wihch
should not be too hard.