change traversal function to only compute the body
This makes the traversal a bit better to understand because it only has
this purpose. Unfortunately it still has many arguments (the name
function is used for mutual recursive calls and the underscore_pattern
is needed for the pattern matches in the branches)
This makes the fixpoint generation functions more similar to the lemma
generation functions because we compute the type separately.