try out traced syntax
There are several possible formalizations.
The easiest to generate code for only saves the distances between
indices in the list.
If we save the indices directly code generation becomes a bit harder.
But both approaches use a recursively defined function.

Using an inductive predicate did not work at all because of the index restriction.