Fix package installation.
We used to copy the static files based on the cwd of the running
program.
This does not work if the executable is installed by opam because cwd is
where the user called the program from.
Instead, we need to install the static files using dune and then
reconstruct the path.